[refactor] : handler라는 명칭 제거
- FoodStorageHandler => FoodStorage
- FoodItemHandler => Food
- TabButtonHandler => FavoriteService
